MFC r258625: sdt: add support for solaris/illumos style DTRACE_PROBE macros
Note: in head KDTRACE_HOOKS is defined in opt_global.h and thus there is
no need to explicitly include any header to check that option.
In this branch we have to include opt_kdtrace.h and, so, this commit
carries additional changes to accomodate for that difference.

Sponsored by:	HybridCluster


git-svn-id: https://svn.freebsd.org/base/stable/9@262059 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f