Create a generic IMX6 kernel config.  This is based on the existing WANDBOARD.common config, but with the freescale-specific optons and devices all together at the bottom now.  In addition to reformatting and shuffling lines around, two new options are added because they're now known to work, VFP and FREEBSD_BOOT_LOADER.
This config does not include any static DTB, it requires that u-boot
provide a DTB (or a custom kernel config can compile one in).

This will supercede all the existing WANDBOARD* configs, but those will
be left around for a while to help people transition their customized
configs to include this new one instead.


git-svn-id: https://svn.freebsd.org/base/head@262354 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f