MFC r262310:
Move the part in bsd.own.mk that sets -Wno-c++11-extensions for clang to
bsd.sys.mk, where it really belongs.  This also causes the flag to get
added when clang is *not* the default system compiler, but is still
used, e.g. by setting WITH_CLANG_IS_CC manually.


git-svn-id: https://svn.freebsd.org/base/stable/9@262459 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f