Increase maximum number of columns to support 1920x1200 displays.
This is a MFC of r262480 and r262502.

Please note that this is changing a constant in a header file so it
could affect ABIs.  However, as far as I can tell, this is apparently
only used in code to declare static arrays and thus it should be safe
to change...

PR:		kern/180558


git-svn-id: https://svn.freebsd.org/base/stable/10@262674 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f