Fix pkill tests so that they run cleanly with prove.
This fixes a pgrep test that assumed that PID 2 was named g_event.  This
does not seem to be the case any longer (and I don't know if it ever was
in all possible setups).

Change this test to use the idle loop instead and determine its expected
PID using ps without assuming any specific ID.


git-svn-id: https://svn.freebsd.org/base/head@262961 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f