MFC 239138 (by andrew):
Import compiler-rt r160957.

This is mostly a no-op other than for ARM where it adds missing
__aeabi_mem* and __aeabi_*divmod functions. Even on ARM these will remain
unused until the rest of the ARM EABI code is merged.

MFC 245628 (by andrew):

Import compiler-rt r172839.

This brings in __aeabi_lcmp and __aeabi_ulcmp. It also fixes the spelling
of __aeabi_f2lz. Both changes originated on the arm_eabi project branch.

MFC 245641:

Add a newline at the end of the file to stop gcc from complaining

MFC 245642 (by andrew):

Don't use the pcs attribute on compilers that don't support it. We can
revert this when we stop supporting old versions of gcc.


git-svn-id: https://svn.freebsd.org/base/stable/9@263560 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f